move test over to correct place
